App/requirements.txt
altgraph==0.17.4
asttokens==3.0.0
autocommand==2.2.2
backports.tarfile==1.2.0
beautifulsoup4==4.11.1
certifi==2021.5.30
chardet==4.0.0
click==8.1.8
colorama==0.4.6
commonmark==0.9.1
debugpy==1.8.11
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
idna==2.10
imap-tools==1.9.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.0.1
json5==0.10.0
jsonpointer==3.0.0
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
MarkupSafe==2.0.1
matplotlib-inline==0.1.7
mistune==3.1.0
more-itertools==10.6.0
mutagen==1.47.0
nest-asyncio==1.6.0
Nuitka==2.6.6
numpy==1.26.4
openpyxl==3.1.2
ordered-set==4.1.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
Pillow==11.0.0
platformdirs==4.3.6
prettytable==3.12.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psutil==6.1.1
pure_eval==0.2.3
pycparser==2.22
pycryptodomex==3.21.0
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyodbc==5.2.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytube==12.0.0
pytz==2024.1
pywin32==309
pywin32-ctypes==0.2.3
pywinpty==2.0.14
pyxlsb==1.0.10
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.25.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.5.1
rpds-py==0.22.3
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
spark-parser==1.8.9
stack-data==0.6.3
tinycss2==1.4.0
tk==0.1.0
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
uncompyle6==3.9.2
uri-template==1.3.0
urllib3==1.26.6
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
widgetsnbextension==4.0.13
win10toast==0.9
win11toast==0.35
winsdk==1.0.0b10
xdis==6.1.3
XlsxWriter==3.2.0
youtube-search==2.1.0
yt-dlp==2022.2.4
zstandard==0.23.0
